% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cart_split}
\alias{cart_split}
\title{cart_split}
\usage{
cart_split(y, x)
}
\arguments{
\item{y}{Outcome variable}

\item{x}{Explanatory variable}
}
\description{
This code is adapted from the vignette in https://github.com/cran/rpart/blob/master/tests/usersplits.R
The only thing we change from that example is the splitting function (stemp), which is the focus of our approach
For more details on the other functions and rpart more generally, check the vignette and package documentation
}
