% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{catt.per.region}
\alias{catt.per.region}
\title{catt.per.region
Obtain the CATT predictions for the main dataset for the points in each PTA region}
\usage{
catt.per.region(beta1, beta0, regions)
}
\arguments{
\item{beta1}{beta_1 prediction from main regression}

\item{beta0}{beta_0 prediction from main regression}

\item{regions}{PTA regions, output from pta.nodes function}
}
\description{
catt.per.region
Obtain the CATT predictions for the main dataset for the points in each PTA region
}
