% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fine_part}
\alias{fine_part}
\title{fine_part}
\usage{
fine_part(list)
}
\arguments{
\item{list}{List that contains all regions where PTA holds, from the most aggregated to most disaggregated ones}
}
\description{
Auxiliary function to obtain the finest partition for which delta1-delta0 ~ 0 in the CART tree
}
