% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_node_paths}
\alias{get_node_paths}
\title{get_node_paths}
\usage{
get_node_paths(rpart_obj)
}
\arguments{
\item{rpart_obj}{Output from the rpart call}
}
\description{
Auxiliary function to determine the paths to all nodes -- internal or leaf -- in CART
We make use of the fact that the nodes are labeled as follows:
parent node i splits into left node 2i  and right node 2i+1
}
