% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{points.per.region}
\alias{points.per.region}
\title{points.per.region}
\usage{
\method{points}{per.region}(listA, listB)
}
\arguments{
\item{listA}{List of finest partitions for which PTA holds}

\item{listB}{List which stores the path that reaches the leaf node of each i}
}
\description{
Auxiliary function to create a matrix with N rows and ncol = number of PTA regions
Each column is a boolean vector of size N equal to TRUE if a point falls into the region
associated to that column
}
