% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pta.nodes}
\alias{pta.nodes}
\title{pta.nodes}
\usage{
pta.nodes(placebo_cart, epsilon)
}
\arguments{
\item{placebo_cart}{CART tree, output from placebo.cart function}

\item{epsilon}{Parameter that dictates how close the trends between two groups needs to be for PTA to be considered valid}
}
\description{
This function extracts the finest partitions of the data which CART flags as PTA regions
and stores which points belong in each region
}
