% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{results.per.region}
\alias{results.per.region}
\title{results.per.region
Output results}
\usage{
results.per.region(
  x,
  delta1_aux,
  delta0_aux,
  delta1_main,
  delta0_main,
  regions,
  CART = NULL
)
}
\arguments{
\item{x}{feature set}

\item{delta1_aux}{delta_1 prediction from auxiliary regression}

\item{delta0_aux}{delta_0 prediction from auxiliary regression}

\item{delta1_main}{delta_1 prediction from main regression}

\item{delta0_main}{delta_0 prediction from main regression}

\item{regions}{PTA regions, output from pta.nodes function}

\item{CART}{rpart object, either NULL, if saveCART==FALSE, or the object, if saveCART==TRUE}
}
\value{
list with: 1- Difference in betas per region; 2- CATT estimates per region; 3- matrix with regions; 4- lists with x per region and x with no flagged region; 6- CART tree
}
\description{
results.per.region
Output results
}
