% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{results}
\alias{results}
\title{results}
\usage{
results(x, gamma1, gamma0, bta1, beta0, placebo_cart, epsilon, saveCART)
}
\arguments{
\item{x}{feature set}

\item{gamma1}{gamma_1 prediction}

\item{gamma0}{gamma_0 prediction}

\item{bta1}{beta_1 + tau_1 + alpha_1 prediction}

\item{beta0}{beta_0 prediction}

\item{placebo_cart}{rpart object}

\item{epsilon}{Parameter that dictates how close the trends between two groups needs to be for PTA to be considered valid}

\item{saveCART}{boolean: should rpart object be stored in the output}
}
\description{
Obtain CATT and delta gamma for each point based on the deepest PTA node they reach
}
