% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{searchPTA}
\alias{searchPTA}
\title{searchPTA}
\usage{
searchPTA(
  x,
  beta1_placebo,
  beta0_placebo,
  beta1_main,
  beta0_main,
  epsilon,
  saveCART = TRUE,
  ...
)
}
\arguments{
\item{x}{Feature set from main data}

\item{epsilon}{Parameter that dictates how close the trends between two groups needs to be for PTA to be considered valid}

\item{saveCART}{Whether or not to save the CART tree fit in the first step}

\item{...}{Additional arguments for rpart.control --- control CART tree growth}

\item{delta1_aux}{delta_1 prediction from auxiliary regression}

\item{delta0_aux}{delta_0 prediction from auxiliary regression}

\item{delta1_main}{delta_1 prediction from main regression}

\item{delta0_main}{delta_0 prediction from main regression}
}
\description{
Wrapper function to perform all operations
}
