% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{placebo.cart}
\alias{placebo.cart}
\title{placebo.cart}
\usage{
placebo.cart(x, gamma1, gamma0, epsilon, wt, ...)
}
\arguments{
\item{x}{Main data feature set}

\item{gamma1}{gamma_1 prediction}

\item{gamma0}{gamma_0 prediction}

\item{epsilon}{Parameter that dictates how close the trends between two groups needs to be for PTA to be considered valid}

\item{wt}{weight vector used for calculating the g=1 and g=0 means}

\item{...}{Additional arguments for rpart fit (see rpart.control)}
}
\description{
Fit CART to placebo (pre-trend) predictions
}
