% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{placebo.cart}
\alias{placebo.cart}
\title{placebo.cart}
\usage{
placebo.cart(x, beta1, beta0, epsilon)
}
\arguments{
\item{x}{Main data feature set}

\item{beta1}{beta_1 prediction for main data from placebo regression}

\item{beta0}{beta_0 prediction for main data from placebo regression}

\item{epsilon}{Parameter that dictates how close the trends between two groups needs to be for PTA to be considered valid}
}
\description{
Fit CART to predictions for main data from placebo regression
}
