% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{searchPTA}
\alias{searchPTA}
\title{searchPTA
Wrapper function to perform all operations}
\usage{
searchPTA(x, beta1_placebo, beta0_placebo, beta1_main, beta0_main, epsilon)
}
\arguments{
\item{x}{Feature set from main data}

\item{beta1_placebo}{beta_1 prediction for main data from placebo regression}

\item{beta0_placebo}{beta_0 prediction for main data from placebo regression}

\item{beta1_main}{beta_1 prediction from main regression}

\item{beta0_main}{beta_0 prediction from main regression}

\item{epsilon}{Parameter that dictates how close the trends between two groups needs to be for PTA to be considered valid}
}
\description{
searchPTA
Wrapper function to perform all operations
}
